Small changes to gen_particle_histories

+++ STFTools.py
@@ -575,7 +575,7 @@ def gen_particle_history_serial(base_halo_data,snaps=[],verbose=1):
 
         ###initialise our new flag arrays
         print('Sorting by IDs ...')
-        Processed_Flags_FRESH=[df(np.column_stack((Particle_IDs_FRESH[itype],list(range(N_Particles_FRESH[itype])),np.zeros(N_Particles_FRESH[itype]),np.zeros(N_Particles_FRESH[itype]))),columns=['ParticleID','ParticleIndex','Processed_L1','Processed_L2']).sort_values(['ParticleID']) for itype in range(len(N_Particles_FRESH))]
+        Processed_Flags_FRESH=[df(np.column_stack((Particle_IDs_FRESH[itype],list(range(N_Particles_FRESH[itype])),np.zeros(N_Particles_FRESH[itype]),np.zeros(N_Particles_FRESH[itype]))),columns=['ParticleID','ParticleIndex','Processed_L1','Processed_L2'],dtype=int).sort_values(['ParticleID']) for itype in range(len(N_Particles_FRESH))]
         print('Finished sorting by IDs')
         
         print(Processed_Flags_FRESH[0])
